Add easygui, give nap capability, make python2
This is just a quick little throw-together so I can sleep/use a timer
form the command line. This is probably going to be overridden
eventually. For now, it is a floating boat made of scrap. But it floats.
I will probably use a different GUI framework as well, so I can keep
python3 abilities.